+++ Ruddarr/Views/Movies/MovieEditView.swift
@@ -31,7 +31,7 @@ struct MovieEditView: View {
                     undoMovieChanges()
                 }
             }
-            .alert(
+            .sensoryAlert(
                 isPresented: instance.movies.errorBinding,
                 error: instance.movies.error
             ) { _ in