+++ Ruddarr/Views/SeriesView.swift
@@ -104,7 +104,7 @@ struct SeriesView: View {
                 guard let searchRequest, await searchRequest.waitForDebounce() else { return }
                 updateDisplayedSeries()
             }
-            .alert(isPresented: $alertPresented, error: error) { _ in
+            .sensoryAlert(isPresented: $alertPresented, error: error) { _ in
                 Button("OK") { error = nil }
             } message: { error in
                 Text(error.recoverySuggestionFallback)
@@ -139,6 +139,7 @@ struct SeriesView: View {
             .buttonStyle(.plain)
             .id(series.id)
         }
+        .animation(.snappy, value: instance.series.cachedItems.map(\.id))
         .viewBottomPadding()
         .scenePadding(.horizontal)
         #if os(iOS)
@@ -227,7 +228,7 @@ struct SeriesView: View {
 
             if Occurrence.since(lastMetadataFetch) > cacheInSeconds {
                 if let model = await instance.fetchMetadata() {
-                    settings.saveInstance(model)
+                    settings.saveInstanceMetadata(model)
                     Occurrence.occurred(lastMetadataFetch)
                 }
             }